Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int n1,n2,n3,avg;
-    scanf("%d%d%d",&n1,&n2,&n3);
+    float n1,n2,n3,avg;
+    scanf("%f%f%f",&n1,&n2,&n3);
     avg=(n1+n2+n3)/3;
-    printf("Average: %.2d",avg);
+    printf("Average: %.2f",avg);
 }
